+++ app/src/main/java/com/techcourse/AppWebApplicationInitializer.java
@@ -2,9 +2,9 @@ package com.techcourse;
 
 import jakarta.servlet.ServletContext;
 import jakarta.servlet.ServletRegistration;
-import nextstep.mvc.AnnotationHandlerAdapter;
+import nextstep.mvc.controller.adapter.AnnotationHandlerAdapter;
 import nextstep.mvc.DispatcherServlet;
-import nextstep.mvc.controller.tobe.AnnotationHandlerMapping;
+import nextstep.mvc.controller.mapping.AnnotationHandlerMapping;
 import nextstep.web.WebApplicationInitializer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
