+++ app/src/main/java/com/techcourse/ManualHandlerMapping.java
@@ -6,9 +6,9 @@ import com.techcourse.controller.manual.LogoutController;
 import com.techcourse.controller.manual.RegisterController;
 import com.techcourse.controller.manual.RegisterViewController;
 import jakarta.servlet.http.HttpServletRequest;
-import nextstep.mvc.HandlerMapping;
 import nextstep.mvc.controller.asis.Controller;
 import nextstep.mvc.controller.asis.ForwardController;
+import nextstep.mvc.controller.mapping.HandlerMapping;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
