+++ mvc/src/main/java/nextstep/mvc/DispatcherServlet.java
@@ -4,6 +4,8 @@ import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import nextstep.mvc.controller.adapter.HandlerAdapter;
+import nextstep.mvc.controller.mapping.HandlerMapping;
 import nextstep.mvc.exception.NotExistException;
 import nextstep.mvc.view.ModelAndView;
 import nextstep.mvc.view.View;
