+++ mvc/src/main/java/nextstep/mvc/controller/mapping/AnnotationHandlerMapping.java
@@ -1,8 +1,7 @@
-package nextstep.mvc.controller.tobe;
+package nextstep.mvc.controller.mapping;
 
 import jakarta.servlet.http.HttpServletRequest;
-import nextstep.mvc.ControllerScanner;
-import nextstep.mvc.HandlerMapping;
+import nextstep.mvc.controller.ControllerScanner;
 import nextstep.web.annotation.RequestMapping;
 import nextstep.web.support.RequestMethod;
 import org.reflections.ReflectionUtils;
